+++ CHANGES.md
@@ -3,6 +3,8 @@
 
 * add `titiler.core` as dependencies to reduce code duplication.
 * update code and templates to follow `titiler.core` specs.
+* refactor dependencies to remove bidx in info/metadata/stats (https://github.com/developmentseed/rio-viz/pull/37)
+* refactor UI (https://github.com/developmentseed/rio-viz/pull/38)
 
 ## 0.6.1 (2021-04-08)